\name{popmcarmatest-package}
\alias{popmcarmatest-package}
\alias{popmcarmatest}
\docType{package}
\title{
\packageTitle{popmcarmatest}
}
\description{
\packageDescription{popmcarmatest}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{popmcarmatest}
\packageIndices{popmcarmatest}
An overview of how to use the package, including the most important functions
}
\author{Juho Kopra
\packageAuthor{popmcarmatest}

Maintainer: Juho Kopra\packageMaintainer{popmcarmatest}
}
\references{
Literature or other references for background information
}
Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory
\keyword{ package }
\seealso{
Optional links to other man pages, e.g.

}
\examples{
simple examples of the most important functions
}
